chore: util 파일 옮김, 필요없는 로직 삭
